fix(gitlab-ci-pipeline-exporter): update serviceAccountName template
The name of the template was changed from `exporter.serviceAccountName`
to `app.serviceAccountName` in acea05b, but not updated in the
Deployment. This caused the templating to fail when `rbac.enabled` was
set.

```
Error: template: gitlab-ci-pipelines-exporter/templates/deployment.yaml:30:38: executing "gitlab-ci-pipelines-exporter/templates/deployment.yaml" at <{{template "exporter.serviceAccountName" .}}>: template "exporter.serviceAccountName" not defined
```